+++ pyproject.toml
@@ -9,6 +9,13 @@ dependencies = [
     "pyspark<4.0",
 ]
 
+[build-system]
+requires = ["setuptools", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
 [project.optional-dependencies]
 dev = [
     "pytest>=8.4.0",
